# api/swagger/swagger.yaml
definitions:
...
  github_com_deva-labs_univia_pkg_types.StatusBadRequest:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Bad Request
        type: string
    type: object
...
  github_com_deva-labs_univia_pkg_types.StatusInternalError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: Internal Server Error
        type: string
    type: object
...
  github_com_deva-labs_univia_pkg_types.StatusUnauthorized:
    properties:
      code:
        example: 401
        type: integer
      message:
        example: Unauthorized
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/posts:
    get:
      description: Get detailed information of a single post by ID
      parameters:
      - description: Post ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get details of this post
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID is required
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusUnauthorized'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusInternalError'
      security:
      - BearerAuth: []
      summary: Get post details
      tags:
      - Social Routes
    put:
      consumes:
      - multipart/form-data
      description: Update post content, categories, and media files
      parameters:
      - description: Post ID
        in: formData
        name: id
        required: true
        type: string
      - description: Post content
        in: formData
        name: content
        required: true
        type: string
      - collectionFormat: csv
        description: List of category UUIDs
        in: formData
        items:
          type: string
        name: category_ids
        required: true
        type: array
      - description: Media files (multiple allowed)
        in: formData
        name: media
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Updated post successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusUnauthorized'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusInternalError'
      security:
      - BearerAuth: []
      summary: Update a post
      tags:
      - Social Routes
...
  /api/v1/posts/create:
    post:
      consumes:
      - multipart/form-data
      description: Create a post with content, categories, and media uploads
      parameters:
      - description: Post content
        in: formData
        name: content
        required: true
        type: string
      - collectionFormat: csv
        description: List of category UUIDs
        in: formData
        items:
          type: string
        name: category_ids
        required: true
        type: array
      - description: Media files (multiple allowed)
        in: formData
        name: media
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Post Created Successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusUnauthorized'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusInternalError'
      security:
      - BearerAuth: []
      summary: Create a new post
      tags:
      - Social Routes
...
swagger: "2.0"